refactor(MapWidgetContainer)
