+++ src/base/static/components/organisms/map-widget-container.tsx
